+++ webpack.common.js
@@ -28,8 +28,8 @@ module.exports = (cssFilename, outputFilename, minimizeCss) => ({
     subscriptionsLandingPageStyles: 'pages/subscriptions-landing/subscriptionsLanding.scss',
     contributionsLandingPage: 'pages/contributions-landing/contributionsLanding.jsx',
     contributionsLandingPageStyles: 'pages/contributions-landing/contributionsLanding.scss',
-    newContributionsLandingPage: 'pages/contributions-landing/newContributionsLanding.jsx',
-    newContributionsLandingPageStyles: 'pages/contributions-landing/newContributionsLanding.scss',
+    newContributionsLandingPage: 'pages/new-contributions-landing/contributionsLanding.jsx',
+    newContributionsLandingPageStyles: 'pages/new-contributions-landing/contributionsLanding.scss',
     newThankYouPage: 'pages/thank-you/newThankYou.jsx',
     digitalSubscriptionLandingPage: 'pages/digital-subscription-landing/digitalSubscriptionLanding.jsx',
     digitalSubscriptionLandingPageStyles: 'pages/digital-subscription-landing/digitalSubscriptionLanding.scss',
